+++ tests/Helpers/TemplaterTest.php
@@ -265,6 +265,35 @@ final class TemplaterTest extends \PHPUnit_Framework_TestCase
                 array(true, 'table'),
                 'SELECT * FROM `table`',
             ),
+            array(
+                'SELECT ?cols FROM `t`',
+                array(array(array('t', 'x'))),
+                'SELECT `t`.`x` FROM `t`',
+            ),
+            array(
+                'SELECT ?col,?col,?col,?c,?col FROM `t`',
+                array(
+                    'a',
+                    array('db' => 'd', 'table' => 't', 'col' => 'c', 'func' => 'COUNT', 'as' => 'zz'),
+                    array('col' => 'cc', 'as' => 'aa'),
+                    array('value' => 3, 'func' => 'COUNT'),
+                    array('value' => 's"s'),
+                ),
+                'SELECT `a`,COUNT(`d`.`t`.`c`) AS `zz`,`cc` AS `aa`,COUNT(3),"s\"s" FROM `t`',
+            ),
+            array(
+                'SELECT ?cols FROM `t`',
+                array(
+                    array(
+                        'a',
+                        array('db' => 'd', 'table' => 't', 'col' => 'c', 'func' => 'COUNT', 'as' => 'zz'),
+                        array('col' => 'cc', 'as' => 'aa'),
+                        array('value' => 3, 'func' => 'COUNT'),
+                        array('value' => 's"s'),
+                    ),
+                ),
+                'SELECT `a`,COUNT(`d`.`t`.`c`) AS `zz`,`cc` AS `aa`,COUNT(3),"s\"s" FROM `t`',
+            ),
         );
     }